+++ CHANGELOG.rst
@@ -6,6 +6,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on `Keep a Changelog <https://keepachangelog.com>`_.
 
+0.4.5
+------
+
+Added
++++++++
+- The new ``AbstractEpistasis.single_mut_effects`` method.
+
+- Options ``returnformat`` and ``stringency_param`` to ``AbstractEpistasis.preferences`` and ``utils.scores_to_prefs``.
+
+Changed
++++++++
+- ``AbstractEpistasis.preferences`` and ``utils.scores_to_prefs`` return site as integer.
+
 0.4.4
 ------